+++ frontend/src/components/camera/CameraView.tsx
@@ -1,110 +1,111 @@
-"use client";
+'use client';
 
-import { useEffect, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
+import React, { useRef, useState, useEffect } from 'react';
+import { useAuth } from '@/lib/AuthContext';
+import { fetcher, API_BASE } from '@/lib/api';
 
 export default function CameraView() {
   const videoRef = useRef<HTMLVideoElement>(null);
   const canvasRef = useRef<HTMLCanvasElement>(null);
-  const router = useRouter();
   const [photoDataUrl, setPhotoDataUrl] = useState<string | null>(null);
+  const [location, setLocation] = useState<{ lat: number; lng: number } | null>(null);
+  const { user } = useAuth();
 
-  // カメラ起動
   useEffect(() => {
+    if (!navigator.mediaDevices || !videoRef.current) return;
+
     navigator.mediaDevices.getUserMedia({ video: true })
       .then((stream) => {
         if (videoRef.current) {
           videoRef.current.srcObject = stream;
-          videoRef.current.play();
         }
       })
-      .catch((err) => {
-        console.error("カメラ起動失敗:", err);
-        alert("カメラの使用が許可されていません。");
-      });
-
-    return () => {
-      if (videoRef.current?.srcObject instanceof MediaStream) {
-        videoRef.current.srcObject.getTracks().forEach(track => track.stop());
-      }
-    };
+      .catch((err) => console.error('カメラ取得失敗:', err));
   }, []);
 
-  // 写真を撮る
-  const capturePhoto = () => {
-    if (!videoRef.current || !canvasRef.current) return;
-
+  const takePhoto = () => {
     const canvas = canvasRef.current;
     const video = videoRef.current;
+    if (!canvas || !video) return;
 
     canvas.width = video.videoWidth;
     canvas.height = video.videoHeight;
+    const ctx = canvas.getContext('2d');
+    if (ctx) {
+      ctx.drawImage(video, 0, 0);
+      const dataUrl = canvas.toDataURL('image/jpeg');
+      setPhotoDataUrl(dataUrl);
+    }
 
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return;
-
-    ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
-    const dataUrl = canvas.toDataURL("image/png");
-    setPhotoDataUrl(dataUrl);
+    // 位置情報取得
+    navigator.geolocation.getCurrentPosition(
+      (pos) => {
+        setLocation({ lat: pos.coords.latitude, lng: pos.coords.longitude });
+      },
+      (err) => {
+        console.error('位置情報取得失敗:', err);
+      }
+    );
   };
 
-  console.log(photoDataUrl);
+  const uploadPhoto = async () => {
+    if (!photoDataUrl || !location || !user) return;
+
+    try {
+      const blob = await (await fetch(photoDataUrl)).blob();
+      const formData = new FormData();
+      formData.append('image', blob, 'photo.jpg');
+
+      const token = localStorage.getItem('jwt_token');
+
+      const uploadRes = await fetch(`${API_BASE}/api/places/upload-image`, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+        body: formData,
+      });
+
+      if (!uploadRes.ok) {
+        const errorText = await uploadRes.text();
+        throw new Error(errorText);
+      }
+
+      const { data } = await uploadRes.json();
+      const imageUrl = data.imageUrl;
+
+      await fetcher('/api/places', {
+        method: 'POST',
+        body: JSON.stringify({
+          name: `自動保存地点-${Date.now()}`,
+          description: 'カメラで撮影されたスポット',
+          category: 'spot',
+          address: '不明',
+          prefecture: '不明',
+          lat: location.lat,
+          lng: location.lng,
+          image_url: imageUrl,
+        }),
+      });
+
+      alert('保存に成功しました！');
+    } catch (err) {
+      console.error('送信エラー:', err);
+      alert('送信に失敗しました');
+    }
+  };
 
   return (
-    <div style={{ position: "relative", height: "100vh", backgroundColor: "#000" }}>
-      {!photoDataUrl ? (
-        <video
-          ref={videoRef}
-          style={{ width: "100%", height: "100%", objectFit: "cover" }}
-          autoPlay
-          playsInline
-          muted
-        />
-      ) : (
-        <img
-          src={photoDataUrl}
-          alt="撮影画像"
-          style={{ width: "100%", height: "100%", objectFit: "contain" }}
-        />
+    <div style={{ textAlign: 'center' }}>
+      <video ref={videoRef} autoPlay playsInline style={{ width: '100%' }} />
+      <canvas ref={canvasRef} style={{ display: 'none' }} />
+      <button onClick={takePhoto} style={{ marginTop: '16px' }}>📸 写真を撮る</button>
+      {photoDataUrl && (
+        <>
+          <img src={photoDataUrl} alt="撮影写真" style={{ width: '100%', marginTop: '16px' }} />
+          <button onClick={uploadPhoto} style={{ marginTop: '12px' }}>⬆ サーバに送信</button>
+        </>
       )}
-
-      <canvas ref={canvasRef} style={{ display: "none" }} />
-
-      <div style={{
-        position: "absolute",
-        bottom: "30px",
-        width: "100%",
-        display: "flex",
-        justifyContent: "space-around",
-      }}>
-        <button
-          onClick={() => router.back()}
-          style={{
-            padding: "12px 24px",
-            fontSize: "16px",
-            borderRadius: "8px",
-            backgroundColor: "#555",
-            color: "white",
-            border: "none",
-            cursor: "pointer",
-          }}
-        >
-          戻る
-        </button>
-
-        <button
-          onClick={capturePhoto}
-          style={{
-            width: "60px",
-            height: "60px",
-            borderRadius: "50%",
-            backgroundColor: "#0070f3",
-            border: "none",
-            cursor: "pointer",
-          }}
-          aria-label="写真を撮る"
-        />
-      </div>
     </div>
   );
 }
