+++ frontend/src/lib/api.ts
@@ -9,7 +9,7 @@
  * @version 2.1.5
  */
 
-const API_BASE =
+export const API_BASE =
   process.env.NEXT_PUBLIC_API_BASE_URL || 'http://localhost:3001';
 
 export const fetcher = async <T = unknown>(
