+++ frontend/src/types/user.ts
@@ -1,7 +1,11 @@
 export type User = {
-  id: string;
-  name: string;
+  id: number;
+  username: string;
   email: string;
-  image: string;
-  isPublic: boolean;
+  password?: string; // ← 必要に応じて optional
+  first_name: string;
+  last_name: string;
+  created_at: string;
+  updated_at: string;
+  public_settings: boolean; // ✅ サーバーと一致
 };
